twice-oumu/requirements.txt
Flask==0.12.2
cf-deployment-tracker==1.0.4
cloudant==2.4.0
line-bot-sdk==1.3.0
watson-developer-cloud==0.26.1
